+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/FileSystemModule.kt
@@ -199,6 +199,10 @@ class FileSystemModule : Module() {
         file.move(destination)
       }
 
+      Function("rename") { file: FileSystemFile, newName: String ->
+        file.rename(newName)
+      }
+
       Property("uri") { file ->
         file.asString()
       }
@@ -290,6 +294,10 @@ class FileSystemModule : Module() {
 
       Function("move") { directory: FileSystemDirectory, destination: FileSystemPath ->
         directory.move(destination)
+      }
+
+      Function("rename") { directory: FileSystemDirectory, newName: String ->
+        directory.rename(newName)
       }
 
       Property("uri") { directory ->