+++ packages/expo-file-system/ios/FileSystemModule.swift
@@ -4,6 +4,10 @@ import ExpoModulesCore
 
 @available(iOS 14, tvOS 14, *)
 public final class FileSystemModule: Module {
+  #if os(iOS) || os(tvOS)
+  private lazy var filePickingHandler = FilePickingHandler(module: self)
+  #endif
+
   var documentDirectory: URL? {
     return appContext?.config.documentDirectory
   }
@@ -101,6 +105,20 @@ public final class FileSystemModule: Module {
       }
       downloadTask.resume()
     }
+
+    AsyncFunction("pickFileAsync") { (initialUri: URL?, mimeType: String?, promise: Promise) in
+      #if os(iOS) || os(tvOS)
+      filePickingHandler.presentDocumentPicker(
+        picker: createFilePicker(initialUri: initialUri, mimeType: mimeType),
+        isDirectory: false,
+        initialUri: initialUri,
+        mimeType: mimeType,
+        promise: promise
+      )
+      #else
+      promise.reject(FeatureNotAvailableOnPlatformException())
+      #endif
+    }.runOnQueue(.main)
 
     Function("info") { (url: URL) in
       let output = PathInfo()
@@ -215,6 +233,10 @@ public final class FileSystemModule: Module {
         try file.move(to: to)
       }
 
+      Function("rename") { (file, newName: String) in
+        try file.rename(newName)
+      }
+
       Property("uri") { file in
         return file.url.absoluteString
       }
@@ -277,6 +299,10 @@ public final class FileSystemModule: Module {
 
       Function("move") { (directory, to: FileSystemPath) in
         try directory.move(to: to)
+      }
+
+      Function("rename") { (directory, newName: String) in
+        try directory.rename(newName)
       }
 
       // this function is internal and will be removed in the future (when returning arrays of shared objects is supported)