+++ packages/expo-router/ios/LinkPreview/LinkPreviewNativeModule.swift
@@ -19,6 +19,10 @@ public class LinkPreviewNativeModule: Module {
         }
       }
 
+      Prop("disableForceFlatten") { (_: NativeLinkPreviewView, _: Bool) in
+        // This prop is used in ExpoShadowNode in order to disable force flattening, when display: contents is used
+      }
+
       Events(
         "onPreviewTapped",
         "onPreviewTappedAnimationCompleted",